cRAWL aNIMATION, powerups glow in darkness, tutorial level started

+++ unity/LD45_mm/Assets/Scripts/Modules/Module.cs
@@ -2,7 +2,7 @@
 public abstract class Module
 {
     public abstract string Name { get; }
-    public abstract int Priority { get; }
+    public abstract float Priority { get; }
     public abstract void ActivateModule();
     public abstract void DeactivateModule();
     public virtual void Default()
